requirements.txt
GitPython
flask
flask_sqlalchemy
flask_login
flask_admin
flask_cors
flask_migrate
wtforms
Werkzeug
pandas
numpy
matplotlib
seaborn
plotly
plotly.express
tensorflow
scikit-learn
celery
redis
dynamic_scheduler
structlog
